+++ codegen/idlparser.py
@@ -216,6 +216,7 @@ class IdlParser:
         elif " or " in name:
             name = name.strip("()")
             names = [self.resolve_type(t).strip("'") for t in name.split(" or ")]
+            names = sorted(set(names))
             return f"'Union[{', '.join(names)}]'"
 
         # Triage
@@ -227,6 +228,8 @@ class IdlParser:
             return "object"  # anything, we ignore this stuff anyway
         elif name in ["OffscreenCanvas"]:
             return "object"
+        elif name in ["PredefinedColorSpace"]:
+            return "str"
         else:
             assert name.startswith("GPU")
             name = name[3:]
@@ -270,9 +273,10 @@ class IdlParser:
                 while not line.startswith("};"):
                     line = self.read_line()
                     lines.append(line)
-                classname = lines[0].split("{")[0].split(":")[0].split()[-1]
+                classname_raw, _, base_raw = lines[0].split("{")[0].partition(":")
+                classname = classname_raw.split()[-1]
                 # Collect base classes
-                based_on = []
+                based_on = list(base_raw.split())
                 while self.peek_line().startswith(classname + " includes "):
                     line = self.read_line()
                     based_on.append(line.split()[-1].rstrip(";"))